Using import/export notation eveyrwhere in backend now, and some cleaning up
