+++ src/node-backend/models/account-models.js
@@ -2,7 +2,7 @@
 
 // TODO should those be classes or objects?
 
-function Credentials(username = "", password = "", email = "") {
+export function Credentials(username = "", password = "", email = "") {
     this.username = username;
     this.password = password;
     this.email = email;
