+++ src/node-backend/package.json
@@ -10,6 +10,7 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "type": "module",
   "dependencies": {
     "express": "^4.18.2",
     "mariadb": "^3.0.1",
