+++ src/node-backend/routes/signup-route.js
@@ -1,14 +1,9 @@
 import express from "express";
-import * as signup from "../controller/signup-controller";
-
-const router = express.Router();
+import * as signupController from "../controller/signup-controller.js";
+export const router = express.Router();
 
 // to test: use postman, select Body -> raw type, json, format: "username":"exampleUser, "password":"examplePassword"
 
 // teacher sign up
-router.post('/teacherSignup', signup.doTeacherSignUp);
-router.post('/teacherEmailConfirmed', signup.confirmTeacherAccount);
-
-// router.post('/createClassroom', createClassroom);
-
-export default router;
+router.post('/teacherSignup', signupController.doTeacherSignUp);
+router.post('/teacherEmailConfirmed', signupController.confirmTeacherAccount);
